pyproject.toml
[project]
name = "image-lora-trainer"
version = "0.1.0"
description = "Train LoRA adapters for Stable Diffusion, SDXL, and Z-Image-Turbo"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.48.2",
    "click>=8.3.1",
    "diffusers @ git+https://github.com/huggingface/diffusers.git",
    "peft>=0.18.0",
    "pillow>=12.0.0",
    "protobuf>=6.33.1",
    "safetensors>=0.4.0",
    "sentencepiece>=0.2.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.3",
    "ultralytics>=8.3.235",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.8",
]
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }
